Rename Components::Bootable to Provider
